Use decimal instead of float/double for non-integer numeric values (#183)
* Use decimal instead of float/double for non-integer numeric values

Replace all double/float types with decimal in protocol types, spec models,
and state records for exact representation of costs, limits, and other
user-facing numbers.

- ProtocolMessages.cs: ActionCommand.Value, ProtocolCostType.DefaultCostLimit,
  ProtocolCostValue.Value, ProtocolConstraint.Value, ProtocolCondition.Value,
  ProtocolRepeat.Value all use decimal
- RosterTypes.cs: CostState.Value uses decimal
- RosterSpecModels.cs: ActionDef.Value and ExpectedCostDef.Value use decimal
- IRosterEngine.cs: SetCostLimit parameter uses decimal
- RosterRunner.cs: simplified AssertEqual for decimal (exact equality)
- Java interop: explicit (double) casts at the boundary where decimal
  values are passed to IKVM Java methods
- NR adapter: NrCostSnapshot.Value uses decimal

Fixes #179


* Use explicit 1m decimal literal for default values

Address review feedback: use 1m instead of 1 for decimal defaults
in ProtocolRepeat.Value and JavaModelFactory.CreateRepeat to make
the intended type explicit and consistent with the decimal migration.


---------

+++ src/BattleScribeSpec.BattleScribe/BattleScribeRosterEngine.cs
@@ -167,7 +167,7 @@ public sealed class BattleScribeRosterEngine : IRosterEngine
             "DuplicateForce is not supported by the BattleScribe Java engine (no public API).");
     }
 
-    public void SetCostLimit(string costTypeId, double value)
+    public void SetCostLimit(string costTypeId, decimal value)
     {
         var costType = Engine.GetCostTypeById(costTypeId)
             ?? throw new InvalidOperationException($"Cost type '{costTypeId}' not found.");
@@ -233,12 +233,12 @@ public sealed class BattleScribeRosterEngine : IRosterEngine
 
         var costs = JavaListToList<net.battlescribe.model.data.Cost>(roster.getCosts());
         var costStates = costs.Select(c =>
-            new CostState(c.getName() ?? "", c.getTypeId() ?? "", c.getValue())).ToList();
+            new CostState(c.getName() ?? "", c.getTypeId() ?? "", (decimal)c.getValue())).ToList();
 
         var rawCostLimits = JavaListToList<net.battlescribe.model.data.Cost>(roster.getCostLimits());
         var costLimitStates = rawCostLimits.Count > 0
             ? rawCostLimits.Select(c =>
-                new CostState(c.getName() ?? "", c.getTypeId() ?? "", c.getValue())).ToList()
+                new CostState(c.getName() ?? "", c.getTypeId() ?? "", (decimal)c.getValue())).ToList()
             : null;
 
         return new RosterState(
@@ -626,7 +626,7 @@ public sealed class BattleScribeRosterEngine : IRosterEngine
             sel.getType(),
             sel.getNumber(),
             hidden,
-            [.. costs.Select(c => new CostState(c.getName() ?? "", c.getTypeId() ?? "", c.getValue()))],
+            [.. costs.Select(c => new CostState(c.getName() ?? "", c.getTypeId() ?? "", (decimal)c.getValue()))],
             [.. children.Select(c => CaptureSelection(c, force))],
             Profiles: [.. profiles.Select(CaptureProfile)],
             Rules: [.. rules.Select(r => new RuleState(r.getName() ?? "", r.getDescription() ?? "", r.isHidden(), r.getPage(),
